[test] move pipelines tests to tests/it
