+++ query/tests/it/pipelines/transforms/mod.rs
@@ -12,21 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod pipe;
-mod pipeline;
-mod pipeline_builder;
-mod pipeline_display;
-mod pipeline_walker;
-mod processor;
-mod processor_empty;
-mod processor_merge;
-mod processor_mixed;
-
-pub use pipe::Pipe;
-pub use pipeline::Pipeline;
-pub use pipeline_builder::PipelineBuilder;
-pub use processor::FormatterSettings;
-pub use processor::Processor;
-pub use processor_empty::EmptyProcessor;
-pub use processor_merge::MergeProcessor;
-pub use processor_mixed::MixedProcessor;
+mod transform_aggregator_final;
+mod transform_aggregator_partial;
+mod transform_expression;
+mod transform_filter;
+mod transform_group_by_final;
+mod transform_group_by_partial;
+mod transform_limit;
+mod transform_limit_by;
+mod transform_projection;
+mod transform_sort;
+mod transform_source;
