+++ query/tests/it/pipelines/transforms/transform_aggregator_partial.rs
@@ -17,59 +17,51 @@ use std::sync::Arc;
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use common_planners::{self};
+use databend_query::pipelines::processors::*;
+use databend_query::pipelines::transforms::*;
 use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::pipelines::processors::*;
-use crate::pipelines::transforms::*;
-
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_transform_projection() -> Result<()> {
+async fn test_transform_partial_aggregator() -> Result<()> {
     let ctx = crate::tests::try_create_context()?;
     let test_source = crate::tests::NumberTestData::create(ctx.clone());
 
+    // sum(number), avg(number)
+    let aggr_exprs = &[sum(col("number")), avg(col("number"))];
+    let aggr_partial = PlanBuilder::create(test_source.number_schema_for_test()?)
+        .aggregate_partial(aggr_exprs, &[])?
+        .build()?;
+
+    // Pipeline.
     let mut pipeline = Pipeline::create(ctx.clone());
-    let source = test_source.number_source_transform_for_test(8)?;
+    let source = test_source.number_source_transform_for_test(200000)?;
+    let source_schema = test_source.number_schema_for_test()?;
     pipeline.add_source(Arc::new(source))?;
 
-    if let PlanNode::Projection(plan) = PlanBuilder::create(test_source.number_schema_for_test()?)
-        .project(&[col("number"), col("number")])?
-        .build()?
-    {
-        pipeline.add_simple_transform(|| {
-            Ok(Box::new(ExpressionTransform::try_create(
-                plan.input.schema(),
-                plan.schema.clone(),
-                plan.expr.clone(),
-            )?))
-        })?;
-        pipeline.add_simple_transform(|| {
-            Ok(Box::new(ProjectionTransform::try_create(
-                plan.input.schema(),
-                plan.schema.clone(),
-                plan.expr.clone(),
-            )?))
-        })?;
-    }
+    pipeline.add_simple_transform(|| {
+        Ok(Box::new(AggregatorPartialTransform::try_create(
+            aggr_partial.schema(),
+            source_schema.clone(),
+            aggr_exprs.to_vec(),
+        )?))
+    })?;
+    pipeline.merge_processor()?;
 
+    // Result.
     let stream = pipeline.execute().await?;
     let result = stream.try_collect::<Vec<_>>().await?;
     let block = &result[0];
     assert_eq!(block.num_columns(), 2);
 
+    // BinaryState
     let expected = vec![
-        "+--------+--------+",
-        "| number | number |",
-        "+--------+--------+",
-        "| 7      | 7      |",
-        "| 6      | 6      |",
-        "| 5      | 5      |",
-        "| 4      | 4      |",
-        "| 3      | 3      |",
-        "| 2      | 2      |",
-        "| 1      | 1      |",
-        "| 0      | 0      |",
-        "+--------+--------+",
+        "+--------------------+----------------------------------+",
+        "| sum(number)        | avg(number)                      |",
+        "+--------------------+----------------------------------+",
+        "| 01604116a804000000 | 604116a804000000400d030000000000 |",
+        "+--------------------+----------------------------------+",
     ];
     common_datablocks::assert_blocks_sorted_eq(expected, result.as_slice());
