+++ query/tests/it/pipelines/transforms/transform_group_by_final.rs
@@ -18,60 +18,74 @@ use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
 use common_planners::{self};
+use databend_query::pipelines::processors::*;
+use databend_query::pipelines::transforms::*;
 use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::pipelines::processors::*;
-use crate::pipelines::transforms::*;
-
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_transform_final_aggregator() -> Result<()> {
+async fn test_transform_final_group_by() -> Result<()> {
     let ctx = crate::tests::try_create_context()?;
     let test_source = crate::tests::NumberTestData::create(ctx.clone());
 
-    // sum(number)+1, avg(number)
+    // sum(number), avg(number)
     let aggr_exprs = &[sum(col("number")), avg(col("number"))];
+
+    let group_exprs = &[col("number")];
     let aggr_partial = PlanBuilder::create(test_source.number_schema_for_test()?)
-        .aggregate_partial(aggr_exprs, &[])?
-        .build()?;
-    let aggr_final = PlanBuilder::create(test_source.number_schema_for_test()?)
-        .aggregate_final(test_source.number_schema_for_test()?, aggr_exprs, &[])?
+        .aggregate_partial(aggr_exprs, group_exprs)?
         .build()?;
 
-    // Pipeline.
-    let source = test_source.number_source_transform_for_test(200000)?;
-    let source_schema = test_source.number_schema_for_test()?;
+    let aggr_final = PlanBuilder::create(test_source.number_schema_for_test()?)
+        .aggregate_final(
+            test_source.number_schema_for_test()?,
+            aggr_exprs,
+            group_exprs,
+        )?
+        .build()?;
 
     let mut pipeline = Pipeline::create(ctx.clone());
+    let source = test_source.number_source_transform_for_test(5)?;
+    let source_schema = test_source.number_schema_for_test()?;
     pipeline.add_source(Arc::new(source))?;
     pipeline.add_simple_transform(|| {
-        Ok(Box::new(AggregatorPartialTransform::try_create(
+        Ok(Box::new(GroupByPartialTransform::create(
             aggr_partial.schema(),
             source_schema.clone(),
             aggr_exprs.to_vec(),
-        )?))
+            group_exprs.to_vec(),
+        )))
     })?;
     pipeline.merge_processor()?;
+
+    let max_block_size = ctx.get_settings().get_max_block_size()? as usize;
     pipeline.add_simple_transform(|| {
-        Ok(Box::new(AggregatorFinalTransform::try_create(
+        Ok(Box::new(GroupByFinalTransform::create(
             aggr_final.schema(),
+            max_block_size,
             source_schema.clone(),
             aggr_exprs.to_vec(),
-        )?))
+            group_exprs.to_vec(),
+        )))
     })?;
 
     // Result.
     let stream = pipeline.execute().await?;
     let result = stream.try_collect::<Vec<_>>().await?;
     let block = &result[0];
-    assert_eq!(block.num_columns(), 2);
+    assert_eq!(block.num_columns(), 3);
 
+    // SELECT SUM(number), AVG(number), number from numbers(5) group by number;
     let expected = vec![
-        "+-------------+-------------+",
-        "| sum(number) | avg(number) |",
-        "+-------------+-------------+",
-        "| 19999900000 | 99999.5     |",
-        "+-------------+-------------+",
+        "+-------------+-------------+--------+",
+        "| sum(number) | avg(number) | number |",
+        "+-------------+-------------+--------+",
+        "| 0           | 0           | 0      |",
+        "| 1           | 1           | 1      |",
+        "| 2           | 2           | 2      |",
+        "| 3           | 3           | 3      |",
+        "| 4           | 4           | 4      |",
+        "+-------------+-------------+--------+",
     ];
     common_datablocks::assert_blocks_sorted_eq(expected, result.as_slice());
